Make test more robust when running example
