+++ tests/test_plugins.py
@@ -5,6 +5,7 @@ from zambeze.orchestration.plugins_message_template_engine import (
 )
 from zambeze.orchestration.message.message_factory import MessageFactory
 from zambeze.orchestration.zambeze_types import MessageType, ActivityType
+from zambeze.orchestration.network import getIP
 
 # Standard imports
 import copy
@@ -20,23 +21,6 @@ import time
 import uuid
 
 logger = logging.getLogger(__name__)
-
-
-def getIP(address_or_hostname: str):
-    # Check if this is an ip address, if not check to see if we can
-    # resolve to an IP address by assuming it is a hostname
-    if re.search("[a-zA-Z]", address_or_hostname):
-        # assuming that because it contains characters it is a hostname
-        try:
-            neighbor_vm_ip = socket.gethostbyname(address_or_hostname)
-        except Exception as e:
-            print(e)
-            raise Exception(
-                "Unable resolze ZAMBEZE_CI_TEST_RSYNC_IP to an IP" " Address"
-            )
-    else:
-        neighbor_vm_ip = address_or_hostname
-    return neighbor_vm_ip
 
 
 @pytest.mark.unit
